requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.7.22
cloudinary==1.36.0
confusable-homoglyphs==3.2.0
Django==4.2.7
django-bootstrap4==23.2
django-registration==3.4
djangorestframework==3.14.0
PyMySQL==1.1.0
python-decouple==3.8
pytz==2023.3.post1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.7
